requirements.txt
numpy 
tensorflow
astroid
blinker
click
config
dill
easyocr
Flask
Flask-Login
Flask-SQLAlchemy
greenlet
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
openai
Pillow
platformdirs
pylint
PyMySQL
python-dotenv
SQLAlchemy
tomlkit
typing_extensions
Werkzeug
wrapt
